requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Chemistry and Drug Discovery
rdkit>=2023.3.1
deepchem>=2.7.0

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter and Development
jupyterlab>=4.0.0
ipykernel>=6.25.0

# Experiment Tracking and Optimization
wandb>=0.15.0
optuna>=3.3.0

# Model Interpretability
shap>=0.42.0

# Code Quality and Testing
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0

# Additional Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional GPU support (uncomment if using CUDA)
# torch-scatter>=2.1.0
# torch-sparse>=0.6.0
# torch-cluster>=1.6.0